InputFilter, OutputFilter support added.
Service class now has two default annotations.

@Secure and @OutputFilter

@Secure
Without any parameters, just handles $request->params->access token and tries to bind
the request with an IUser with access token information. It does not do
any authorization though. For that you have to use @Secure annotation
parameters -- which is not decided yet.

@OutputFilter AutoFormat
Depending on the $request->params->type it encodes the output. At the
moment it support **json** and **raw**. If the type is json and
$request->params->callback is set, then the JSON response is wrapped
with a function call.