AutoFormat, Discover and API generation
AutoFormat decodes request body if the Content-Type header of the
request is application/json and merges the decoded parameters with
Request->params.

Base service class now has "@InputFilter AutoFormat" and "@OutputFilter
AutoFormat" instead of just "@OutputFilter AutoFormat" because
AutoFormat now handles request bodies and needs to be processed before
the service too. We could've used @Filter too as it both does pre and
post processing, then, it'd be a bit harder to just disable OutputFilter
or InputFilter.

Discover algorithm is fixed to detect all the services.

Zita::apigen service call is introduced, it generates Javascript API for
the web services.